+++ examples/v2-forum/package.json
@@ -13,6 +13,8 @@
     "@canvas-js/chain-ethereum": "0.6.0-alpha7",
     "@canvas-js/hooks": "0.6.0-alpha7",
     "@canvas-js/templates": "0.6.0-alpha7",
+    "@headlessui/react": "^1.7.17",
+    "@heroicons/react": "^2.0.18",
     "ethers": "^6.8.0",
     "moment": "^2.29.4",
     "react": "^18.2.0",
